refactor(tcf/ui): extract userReadable function
