+++ components/tcf/ui/src/SecondLayer/components/tcf-secondLayer-vendor-expandedContent/index.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import userReadable from '../../services/secondsToUserReadable'
 
 export default function TcfSecondLayerVendorExpandedContent({
   info,
@@ -25,36 +26,6 @@ export default function TcfSecondLayerVendorExpandedContent({
     </>
   )
 
-  const secondsToTimeUnits = totalSeconds => {
-    const result = {}
-    let remaining = totalSeconds
-    const divideTimeUnits = (key, divider) => {
-      const wholeTimeUnit = Math.floor(remaining / divider)
-      if (wholeTimeUnit) {
-        result[key] = wholeTimeUnit
-      }
-      remaining = remaining % divider
-    }
-    divideTimeUnits('days', 86400)
-    divideTimeUnits('hours', 3600)
-    divideTimeUnits('minutes', 60)
-    result.seconds = remaining
-    return result
-  }
-
-  const userReadableTimeUnits = seconds => {
-    if (seconds <= 0) {
-      return i18n.VENDOR_PAGE.GROUPS.EXPANDED.COOKIES.NEGATIVE_OR_ZERO_MAX_AGE
-    }
-    const time = secondsToTimeUnits(seconds)
-    let s
-    s = time.days ? `${time.days} ${i18n.DAYS} ` : ''
-    s += time.hours ? `${time.hours} ${i18n.HOURS} ` : ''
-    s += time.minutes ? `${time.minutes} ${i18n.MINUTES} ` : ''
-    s += time.seconds ? `${time.seconds} ${i18n.SECONDS}` : ''
-    return s
-  }
-
   const CookieAgeInfo = () => {
     return (
       <>
@@ -63,7 +34,7 @@ export default function TcfSecondLayerVendorExpandedContent({
         </h6>
         <p>
           {i18n.VENDOR_PAGE.GROUPS.EXPANDED.COOKIES.COOKIES_MAX_AGE_SECONDS}:{' '}
-          {userReadableTimeUnits(info.cookieMaxAgeSeconds)}
+          {userReadable({seconds: info.cookieMaxAgeSeconds, i18n})}
         </p>
         <p>
           {i18n.VENDOR_PAGE.GROUPS.EXPANDED.COOKIES.USES_NON_COOKIE_ACCESS}:{' '}
